chore: migrate E2EEncryptionSecurityView to hooks (#5085)
* chore: migrate E2EEncryptionSecurityView to hooks

* minor tweak
